1. Integrasi ke smartmining Production, semua module mining,barging dan shipping

+++ MCSWebApp/Controllers/API/Mining/WasteRemovalController.cs
@@ -244,7 +244,7 @@ namespace MCSWebApp.Controllers.API.Mining
                         record.owner_id = CurrentUserContext.AppUserId;
                         record.organization_id = CurrentUserContext.OrganizationId;
                         //record.business_unit_id = HttpContext.Session.GetString("BUSINESS_UNIT_ID");
-
+                        record.integration_status = "NOT APPROVED";
                         #region Validation
 
                         // Source location != destination location
@@ -352,6 +352,8 @@ namespace MCSWebApp.Controllers.API.Mining
                         if (await mcsContext.CanUpdate(dbContext, record.id, CurrentUserContext.AppUserId)
                             || CurrentUserContext.IsSysAdmin)
                         {
+                            if (record.integration_status != "NOT APPROVED")
+                                throw new Exception("Cannot edit data,Please Request for Unapproval"); 
                             var e = new entity();
                             e.InjectFrom(record);
 
@@ -523,6 +525,8 @@ namespace MCSWebApp.Controllers.API.Mining
                     if (await mcsContext.CanDelete(dbContext, key, CurrentUserContext.AppUserId)
                         || CurrentUserContext.IsSysAdmin)
                     {
+                        if (record.integration_status != "NOT APPROVED")
+                            throw new Exception("Cannot edit data,Please Request for Unapproval");
                         dbContext.waste_removal.Remove(record);
                         await dbContext.SaveChangesAsync();
                     }
@@ -568,6 +572,68 @@ namespace MCSWebApp.Controllers.API.Mining
                             }
                         }
 
+                        await tx.CommitAsync();
+
+                        result.Success = true;
+                        return Ok(result);
+                    }
+                    else
+                    {
+                        result.Message = "Invalid data.";
+                    }
+                }
+                catch (Exception ex)
+                {
+                    await tx.RollbackAsync();
+                    logger.Error(ex.ToString());
+                    result.Message = ex.Message;
+                }
+            }
+
+            return new JsonResult(result);
+        }
+        [HttpPut("ApproveUnapprove")]
+        [ApiExplorerSettings(IgnoreApi = true)]
+        public async Task<IActionResult> ApproveUnapprove([FromBody] dynamic Data)
+        {
+            var result = new StandardResult();
+            using (var tx = await dbContext.Database.BeginTransactionAsync())
+            {
+                try
+                {
+                    if (Data != null && Data.selectedIds != null)
+                    {
+                        var selectedIds = ((string)Data.selectedIds)
+                            .Split(",".ToCharArray(), StringSplitOptions.RemoveEmptyEntries)
+                            .ToList();
+                        var waste_removal = await dbContext.waste_removal.Where(o=>selectedIds.Contains(o.id)).ToListAsync();
+                        await _semaphore.WaitAsync();
+                        try
+                        {
+                            foreach (var record in waste_removal)
+                            {
+                                switch (record.integration_status)
+                                {
+                                    case "NOT APPROVED":
+                                        record.integration_status = "REQUESTED FOR APPROVAL";
+                                        break;
+                                    case "REQUESTED FOR APPROVAL":
+                                        record.integration_status = "NOT APPROVED";
+                                        break;
+                                    case "APPROVED":
+                                        record.integration_status = "REQUESTED FOR UNAPPROVAL";
+                                        break;
+                                    case "REQUESTED FOR UNAPPROVAL":
+                                        record.integration_status = "APPROVED";
+                                        break;
+                                }
+                            }
+                        }
+                        finally
+                        {
+                            _semaphore.Release();
+                        }
+                        await dbContext.SaveChangesAsync();
                         await tx.CommitAsync();
 
                         result.Success = true;
@@ -1137,7 +1203,7 @@ namespace MCSWebApp.Controllers.API.Mining
                         sheet = wb.GetSheetAt(0); //get first sheet from workbook
                     }
 
-                    string teks = "";
+                    string teks = $"";
                     bool gagal = false; string errormessage = "";
                     int totalRows = sheet.LastRowNum - sheet.FirstRowNum;
                     using var transaction = await dbContext.Database.BeginTransactionAsync();
@@ -1318,6 +1384,8 @@ namespace MCSWebApp.Controllers.API.Mining
                                 .FirstOrDefault();
                             if (record != null)
                             {
+                                if(record.integration_status != "NOT APPROVED")
+                                   throw new Exception ("Error in Line : " + (i + 1) + " ==> Can't Process data. Please Unapprove the data. Transaction Number: " + record.transaction_number + Environment.NewLine);
                                 var e = new entity();
                                 e.InjectFrom(record);
 
@@ -1342,7 +1410,6 @@ namespace MCSWebApp.Controllers.API.Mining
                                 record.pic = employee_id;
                                 record.advance_contract_id = advance_contract_id;
                                 record.business_unit_id = business_unit_id;
-
                                 await dbContext.SaveChangesAsync();
                             }
                             else
@@ -1359,6 +1426,7 @@ namespace MCSWebApp.Controllers.API.Mining
                                 record.entity_id = null;
                                 record.owner_id = CurrentUserContext.AppUserId;
                                 record.organization_id = CurrentUserContext.OrganizationId;
+                                record.integration_status = "NOT APPROVED";
                                 record.business_unit_id = business_unit_id;
 
                                 record.transaction_number = TransactionNumber;
